requirements.txt
langchain==0.3.1
langchain-anthropic==0.1.22
langchain-community==0.2.11
langchain-core==0.3.6
langchain-openai==0.2.0
langchain-pinecone==0.1.3
langchain-text-splitters==0.3.0
langchainhub==0.1.21
matplotlib==3.9.2
numpy==1.26.3
openai==1.48.0
pandas==2.2.3
pinecone-client==5.0.1
python-dotenv==0.13.0
scikit-learn==1.5.2
tiktoken==0.7.0
umap-learn==0.5.5
